pass string to * transform test

+++ src/main/kotlin/com/hagoapp/f2t/compare/transform/String2FloatTransformer.kt
@@ -24,7 +24,7 @@ class String2FloatTransformer : TypedColumnTransformer {
         }
         return when (dbColumnDefinition.dataType) {
             FLOAT -> src?.toFloat()
-            SMALLINT -> src?.toDouble()
+            DOUBLE -> src?.toDouble()
             DECIMAL -> src?.toBigDecimal()
             else -> null
         }
